Operate on edits rather than a diff
This avoids needing to compute a diff as well as greatly simplifying
cursor positioning (this keeps the cursor position unchanged as edits
are cleanly applied around it).

This commit also drops the code borrowed from the vscode-python
extension and the diff-match-patch dependency, both of which were
only needed to process diffs.